refactor: ChatService 메서드의 sendNotification 메서드를 NotificationService로 이동
