+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/service/chat/ChatService.java
@@ -15,9 +15,6 @@ public interface ChatService {
 
     Slice<ChatRoomRespDto> getChatRooms(Long memberId, Pageable pageable);
 
-    void exitChatRoom(Long memberId, Long chatRoomId);
-
-    //    Slice<ChatMessageListRespDto> getChatMessages(Long memberId, Long chatRoomId, Pageable pageable);
     ChatMessageListRespDto getChatMessages(Long memberId, Long chatRoomId, String cursor, Pageable pageable);
 
     void sendMessage(Message message, Long senderId);
